+++ lib/recourse/helpers/actions.rb
@@ -81,11 +81,11 @@ module Recourse
       def destroy_button(record)
         path = url_for controller: "/#{destroy_action_path}", action: :destroy, id: record
 
-        button_to icon_tag(ICONS[:destroy], class: 'fg-danger'), path,
-                  method: :delete, class: 'btn btn-sm btn-link btn-icon p-0',
-                  aria: { label: t('recourse.delete', model: resource_name) },
-                  form: { data: { turbo_frame: '_top',
-                                  turbo_confirm: destroy_warning(record), } }
+        bh_confirm_button icon_tag(ICONS[:destroy], class: 'fg-danger'), path,
+                          confirm: destroy_warning(record), method: :delete,
+                          class: 'btn btn-sm btn-link btn-icon p-0',
+                          aria: { label: t('recourse.delete', model: resource_name) },
+                          form: { data: { turbo_frame: '_top' } }
       end
     end
   end